Add Panoptes launcher binary.
As to simplify using Panoptes, this commit adds a small launcher which
injects libpanoptes.so into the LD_PRELOAD environment variable.

Additionally, the launcher adds PANOPTES_TOOL=MEMCHECK to the
environment.  In the future, other tools will be supported via the same
libpanoptes.so interposition library.  This flag will allow these tools
to be selectable at runtime.

Sample usage:
    panoptes [--tool=memcheck] my_cuda_program [args]...